+++ src/components/author/AuthorItems.jsx
@@ -29,13 +29,13 @@ const AuthorItems = ( { collection, image, authorID, windowWidth } ) => {
               <button>Buy Now</button>
               <div className="nft__item_share">
                 <h4>Share</h4>
-                <a href="" target="_blank" rel="noreferrer">
+                <a href="https://www.facebook.com" target="_blank" rel="noreferrer">
                   <i className="fa fa-facebook fa-lg"></i>
                 </a>
-                <a href="" target="_blank" rel="noreferrer">
+                <a href="https://www.x.com" target="_blank" rel="noreferrer">
                   <i className="fa fa-twitter fa-lg"></i>
                 </a>
-                <a href="">
+                <a href="mailto:recipient@example.com">
                   <i className="fa fa-envelope fa-lg"></i>
                 </a>
               </div>
